+++ packages/lists/CHANGELOG.md
@@ -7,6 +7,7 @@
 -   `getSelectListWithValues` was renamed to `getSelectListWithKeys`.
 -   `getSelectListWithLabels` was renamed to `getSelectListWithValues`.
 -   `getListValue.value` was renamed to `getListValue.key`.
+-   Added `getListKey` to return the key from a list, based on a value.
 
 ## v1.0.0-beta.1